cli/owners: Consolidate owners logic in utils dir
Move the main codeowners logic into the dedicated package. Only minor
changes here for function signatures to allow the package to not worry
about logging internally but instead return owners errors up to the
caller for handling.
